Merge branch 'Randi' of https://github.com/ar4757/Handwriting-Calculator into Randi
Redesigned calculator activity with the drawing area being a handwritingView with a gesture overlay. This is defined in the xml file.

All gestures are pushed onto the stack and drawn to the underlying view. If there is no overlapping gesture, the last two gestures are popped off the stack and replace with the combined gesture.

EXISTING ISSUE - having to do with the timer. An extra call to recognize gesture is being made when gestures are drawn quickly. This results in an extra action string added to the math expression. This still needs to be debugged.