Merge branch 'Randi' of https://github.com/ar4757/Handwriting-Calculator into Randi
Redesigned calculator activity with the drawing area being a handwritingView with a gesture overlay. This is defined in the xml file.

All gestures are pushed onto the stack and drawn to the underlying view. If there is no overlapping gesture, the last two gestures are popped off the stack and replace with the combined gesture.

FIxed issue with countdown timer causing recognizeGesture to be called an extra time when entering multiple non-overlapping gestures quickly.